+++ CHANGELOG.md
@@ -5,6 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 17-03-2017 Geraint Edwards
 * Fix for save button not appearing
 * Week count in monthly repeatihg event editing now uses ordinals and reserves when counting back from the end of the month
+15-03-2017 Tony Partridge
+* Change from Editor to Creator for Save / Apply on new events
+* Imrpoved canPublishOwnEvents() method to return $canPublishOwn value if set and no conditions met.
+
+14-03-2017 Tony Partridge
+* Added new Custom CSS View and JForm based form
 
 14-03-2017 Geraint Edwards
 * JEvents 3.4.31